# Conflicts:
#	package.json
#	src/source/geojson_source.js
#	src/source/geojson_worker_source.js

+++ src/shaders/fill.fragment.glsl
@@ -1,8 +1,8 @@
-#pragma mapbox: define lowp vec4 color
+#pragma mapbox: define highp vec4 color
 #pragma mapbox: define lowp float opacity
 
 void main() {
-    #pragma mapbox: initialize lowp vec4 color
+    #pragma mapbox: initialize highp vec4 color
     #pragma mapbox: initialize lowp float opacity
 
     gl_FragColor = color * opacity;
