+++ src/shaders/fill_outline.vertex.glsl
@@ -5,11 +5,11 @@ uniform vec2 u_world;
 
 varying vec2 v_pos;
 
-#pragma mapbox: define lowp vec4 outline_color
+#pragma mapbox: define highp vec4 outline_color
 #pragma mapbox: define lowp float opacity
 
 void main() {
-    #pragma mapbox: initialize lowp vec4 outline_color
+    #pragma mapbox: initialize highp vec4 outline_color
     #pragma mapbox: initialize lowp float opacity
 
     gl_Position = u_matrix * vec4(a_pos, 0, 1);
